28.06.2022

+++ bindings/cuda_utils/include/cuda_utils.h
@@ -47,8 +47,8 @@
 
 struct Layer {
     std::string name;
-    std::vector<double> weights;
-    std::vector<double> bias;
+    std::vector<long double> weights;
+    std::vector<long double> bias;
 };
 
 class cudaOnTarget {
